requirements.txt
asgiref==3.5.2
certifi==2022.6.15.1
charset-normalizer==2.1.1
dj-database-url==1.0.0
Django==4.1
django-cors-headers==3.13.0
django-heroku==0.3.1
django-restframework==0.0.1
djangorestframework==3.13.1
gunicorn==20.1.0
heroku==0.1.4
idna==3.3
psycopg2==2.9.3
python-dateutil==1.5
pytz==2022.2.1
requests==2.28.1
sqlparse==0.4.2
tzdata==2022.2
urllib3==1.26.12
whitenoise==6.2.0
